Add a basic GUI to play music in the library.
The GUI is based on GTK+ and plays music using the GStreamer library. It is
currently incredibly basic. It allows the user to add new directories to the
library, but there is currently no way to remove them.

The player offers no playlist functionality, instead constantly shuffling all
music in the library.

Signed-off-by: Jason Lynch <jason@calindora.com>