Allow the user to enqueue their choice of four different tracks.
Most current audio players expect users to manually rate their music (either on
a like/dislike scale or perhaps with 1-5 stars, or something similar. A few
attempt to automatically determine the ratings based on user behavior.

All of these are problematic in that they attempt to assign cardinal ratings to
things, a task which humans can find difficult. One thing that may be easier for
some people, however, is to simply choose which track they like best or most
want to listen to from a small set of choices. This is the beginning of
attempting to implement such a thing. Note that it is currently only a queueing
system, and nothing is actually done with the data. In fact, no data is even
currently saved regarding the selection.

Signed-off-by: Jason Lynch <jason@calindora.com>