Add the ability to preview tracks in the choices list.
Choosing to preview a track pauses playback and instead plays the track as a
preview, to help the user make their selection in case they don't recognize or
remember one or more of the tracks.

The user can switch among all four tracks by clicking the appropriate preview
button, and disable the preview by clicking the currently active preview button.
The idea is to be able to have this preview not affect any future gathering of
statistics the application may do. If the previewed track ends normally, the
application will resume the original track where it was left off as if the user
manually ended the preview.

Signed-off-by: Jason Lynch <jason@calindora.com>