Group files by musicbrainz_trackid into a single track object.
This allows us to collect statistics for recordings as defined by Musicbrainz
rather than for files, preventing one source of duplicates.

It currently applies no logic toward choosing the best file to play, however. In
addition, because it relies on Musicbrainz tags, it requires the files to be
tagged in such a fashion.

Signed-off-by: Jason Lynch <jason@calindora.com>