Change the way scores are calculated.
Previously, the score was an integer that represented the difference
between the number of times it has won and the times it has lost against
a given opposing track. The new score instead represents an estimate of
the probability that the given track would be chosen over the opposing
track.

Signed-off-by: Jason Lynch <jason@calindora.com>