Update to the latest state of the art library.
This does add some minimal debug output, but unless I figure out a way
to make the library a lot faster, I'm probably not removing it.

I think this commit does also add some dead code, but I don't currently
want to discard any of it, as it may be useful for future updates.

Signed-off-by: Jason Lynch <jason@calindora.com>