Simplify the code for the ranking system.
After some experimentation, this code is still much too slow (and
requires way too many comparisons) for anything other than smaller
libraries. As such, it will most likely be replaced with an entirely
different type of algorithm in the next commit, but this will retain the
existing state for posterity.

Signed-off-by: Jason Lynch <jason@calindora.com>