Significantly rewrite the way ratings are determined.
Ratings are now adjusted for each track involved in a comparison after
the comparison according to the Glicko algorithm. This should require on
the order of n log n comparisons to generate useful ratings (similar to
the number of comparisons needed under the old sort-based system), but
it has the advantage of being able to generate ratings at all times,
rather than failing until a full ordering is established.

The rating algorithm unfortunately does not take into account changes in
opponent strength to recalculate ratings, so a track that doesn't see
many comparisons won't have its rating adjusted to take into account
additional information about the tracks it did defeat or lost to. To
counter this, two experimental rating-generation algorithms are also
included in the code in case they prove useful. They are very rough and
are probably not even bug-free, but are included for the sake of
experimentation.

Unfortunately, these changes require essentially nuking the database and
starting over. In theory, it should convert an existing library, but any
track comparison data will be deleted. (In theory, it would have been
possible to port this perhaps, but it wasn't worth it for me.)

Signed-off-by: Jason Lynch <jason@calindora.com>